refactor(store): refactor thread store to encompass both threads and comments
