+++ store.js
@@ -20,7 +20,7 @@ import * as avatarsActions from './store/actions/avatars'
 import * as charactersActions from './store/actions/characters'
 import * as groupsActions from './store/actions/groups'
 import * as rulesetsActions from './store/actions/rulesets'
-import * as threadsActions from './store/actions/threads'
+import * as forumsActions from './store/actions/forums'
 import * as usersActions from './store/actions/users'
 
 
@@ -33,7 +33,7 @@ export const actions = {
   ...charactersActions,
   ...groupsActions,
   ...rulesetsActions,
-  ...threadsActions,
+  ...forumsActions,
   ...usersActions,
 }
