+++ store/reducers/index.js
@@ -4,7 +4,7 @@ import avatars from './avatars'
 import characters from './characters'
 import groups from './groups'
 import rulesets from './rulesets'
-import threads from './threads'
+import forums from './forums'
 import users from './users'
 
 
@@ -15,8 +15,8 @@ export default combineReducers({
   authentication,
   avatars,
   characters,
+  forums,
   groups,
   rulesets,
-  threads,
   users,
 })
